requirement.txt
awesome-slugify==1.6.5
chardet==2.3.0
cssselect==0.9.2
docopt==0.6.2
html2text==2016.5.29
lxml==3.6.4
peewee==2.8.3
pocket-api==0.1.3
prompt-toolkit==1.0.7
py-dateutil==2.2
PyYAML==3.12
readability-lxml==0.6.2
regex==2016.8.27
requests==2.11.1
six==1.10.0
Unidecode==0.4.19
wcwidth==0.1.7
